+++ packages/deepsea-tools/src/utils/getRandomPhone.ts
@@ -1,4 +1,4 @@
-import { digits } from "@constants/index"
+import { digits } from "@/constants/index"
 import { getRandomItemFromArray } from "./getRandomItemFromArray"
 
 /**