+++ packages/deepsea-tools/src/utils/isLegalId.ts
@@ -1,4 +1,4 @@
-import { idReg } from "@constants/index"
+import { idReg } from "@/constants/index"
 import { getIdChecksum } from "./getIdChecksum"
 
 /**