+++ packages/soda-hooks/src/hooks/useAsync.ts
@@ -1,6 +1,6 @@
 import { DependencyList, useEffect } from "react"
 
-import { isAsyncGenerator } from "@utils/isAsyncGenerator"
+import { isAsyncGenerator } from "@/utils/isAsyncGenerator"
 
 export function useAsync(effect: () => AsyncGenerator<void, void, void> | Promise<void>, deps?: DependencyList): void
 export function useAsync(effect: () => AsyncGenerator<void, void, void> | Promise<void>, callback: () => void, deps?: DependencyList): void