+++ packages/soda-next/src/index.ts
@@ -1,1 +1,2 @@
-export * from "@hooks/useQueryState"
+export * from "@/components/Navigate"
+export * from "@/hooks/useQueryState"