+++ packages/soda-react-router/src/index.ts
@@ -1,1 +1,1 @@
-export * from "@hooks/useQueryState"
+export * from "@/hooks/useQueryState"