+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # deepsea-components
 
+## 5.7.11
+
+### Patch Changes
+
+-   更新路径别名，统一使用 @ 符号引入模块
+-   Updated dependencies
+    -   deepsea-tools@5.15.2
+
 ## 5.7.10
 
 ### Patch Changes