+++ packages/soda-next/CHANGELOG.md
@@ -1,5 +1,14 @@
 # soda-next
 
+## 0.3.1
+
+### Patch Changes
+
+-   更新路径别名，统一使用 @ 符号引入模块
+-   Updated dependencies
+    -   deepsea-tools@5.15.2
+    -   soda-hooks@6.1.8
+
 ## 0.3.0
 
 ### Minor Changes